move code to new concern
